backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-allauth
requests
PyJWT
cryptography
djangorestframework-simplejwt[blacklist]
dj-rest-auth[social]>=2.2.0
# Testing
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
factory-boy>=3.3.0
stripe